Review

+++ include/common/protocols/json_rpc/protocol_client.h
@@ -29,7 +29,7 @@ namespace protocols {
 namespace json_rpc {
 
 typedef uint32_t protocoled_size_t;  // sizeof 4 byte
-enum : protocoled_size_t { MAX_COMMAND_LENGTH = UINT32_MAX };
+enum : protocoled_size_t { MAX_COMMAND_LENGTH = UINT16_MAX };
 COMPILE_ASSERT(std::numeric_limits<protocoled_size_t>::max() >= MAX_COMMAND_LENGTH,
                "Protocoled packet size should be greater MAX_COMMAND_LENGTH");
